feat: load display options from cookies in display.js
