+++ src/static/timetracker/js/utils.js
@@ -1,4 +1,4 @@
-export { format_mins, duration, string_to_mins, dayOfWeek, getCookies, parseElementApiInfo, getAllDays }
+export { format_mins, duration, string_to_mins, dayOfWeek, parseElementApiInfo, getAllDays }
 
 function string_to_mins(s) {
   let parts = s.split(":");
@@ -28,19 +28,6 @@ function dayOfWeek(date) {
   let dateObject = new Date(date);
   const days = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
   return days[dateObject.getDay()];
-}
-
-function getCookies() {
-  const cookies = {};
-  const cookieString = document.cookie;
-  if (cookieString) {
-    const cookieArray = cookieString.split('; ');
-    for (let i = 0; i < cookieArray.length; i++) {
-      const cookie = cookieArray[i].split('=');
-      cookies[decodeURIComponent(cookie[0])] = decodeURIComponent(cookie[1]);
-    }
-  }
-  return cookies;
 }
 
 function parseElementApiInfo(element) {
